+++ beacon-clients/beacon-web/src/main.tsx
@@ -3,6 +3,7 @@ import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router-dom";
 
 import App from "./App";
+import { routerBasename } from "./lib/base-path";
 import { BeaconProvider } from "./lib/beacon-context";
 import { AppQueryProvider } from "./lib/query-provider";
 import { ThemeProvider } from "./lib/theme";
@@ -11,7 +12,7 @@ import "./index.css";
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ThemeProvider>
-      <BrowserRouter basename={import.meta.env.BASE_URL.replace(/\/$/, "") || "/"}>
+      <BrowserRouter basename={routerBasename()}>
         <BeaconProvider>
           <AppQueryProvider>
             <App />